+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.7 (2019-01-18)
 
+* Add `publisher` data to the `generate-data-file` tool.
 * Add `repository_user` and `repository_user_page` properties to `ExtensionVersion`, to return user or organization to which the extension's repository belongs.
 
 ## 0.0.6 (2018-11-20)